+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Select: Fixed multiselect operation
 - Select: Fixed handling of 0/false/null values
+- Select: Fixed dropdown list positioning
 
 ## [1.155.0]