+++ CHANGELOG.md
@@ -13,8 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Select: Fixed multiselect operation
-- Select: Fixed handling of 0/false/null values
+- Select: Fixed handling of 0/false values
 - Select: Fixed dropdown list positioning
+- Select: Fixed selection of items with arrows
 
 ## [1.155.0]